# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Denoise X API
  description: |
    Clinical-grade AI X-ray enhancement platform.
    All `/api/v1/*` endpoints return the standard envelope `{ success, data, meta, trace_id }`.
    Errors return `{ code, message, trace_id, status }` with a stable machine-readable
    `code` — the frontend switches on `code`, never on HTTP status text.
  version: 1.0.0
  contact:
    name: Denoise X
servers:
  - url: http://localhost:8000
    description: Local development (gateway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: httpOnly, Secure, SameSite=Lax cookie set on login/refresh
  schemas:
    TraceId:
      type: string
      description: W3C trace id correlating this request across API, RabbitMQ, worker, S3 and DB.
    Envelope:
      type: object
      required: [success, trace_id]
      properties:
        success:
          type: boolean
        data:
          description: Payload. Absent on errors.
        meta:
          type: object
          description: Pagination / request metadata.
        trace_id:
          $ref: "#/components/schemas/TraceId"
      example:
        success: true
        data: {}
        meta: {}
        trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
    ErrorBody:
      type: object
      required: [code, message, trace_id, status]
      properties:
        code:
          type: string
          description: Stable machine-readable error code (see ErrorCode enum).
        message:
          type: string
        trace_id:
          $ref: "#/components/schemas/TraceId"
        status:
          type: integer
          description: HTTP status for convenience.
      example:
        code: SCAN_TOO_LARGE
        message: Maximum upload size is 50 MB.
        trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
        status: 413
...
    ScanStatus:
      type: string
      enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED]
    OutputType:
      type: string
      enum: [ORIGINAL, NOISE_MAP, UNET, ENHANCED]
    ScanSummary:
      type: object
      required: [id, status, original_name, format, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/ScanStatus" }
        original_name: { type: string }
        format:
          type: string
          enum: [PNG, JPEG, DICOM]
        size_bytes: { type: integer }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
        noise_variance: { type: [number, "null"] }
        routing_message: { type: [string, "null"] }
        was_bypassed: { type: [boolean, "null"] }
        processing_time_ms: { type: [number, "null"] }
        model:
          type: [string, "null"]
          description: "model_name@model_version, e.g. n2n_unet@v1.0.0"
        created_at: { type: string, format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
    ScanDetail:
      allOf:
        - $ref: "#/components/schemas/ScanSummary"
        - type: object
          properties:
            outputs:
              type: array
              description: Present once COMPLETED.
              items:
                type: object
                properties:
                  type: { $ref: "#/components/schemas/OutputType" }
                  size_bytes: { type: integer }
                  mime_type: { type: string }
...
    PageMeta:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
  responses:
    Error:
      description: Standard error envelope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    RateLimited:
      description: RATE_LIMITED — Retry-After header provided.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/scans:
    post:
      tags: [Scans]
      operationId: createScan
      summary: Upload an X-ray (PNG/JPEG/DICOM) and enqueue async denoising.
      description: |
        Validates magic bytes, decode, dimensions and 50MB cap, uploads the original
        to object storage, records scan+job rows and publishes `commands.inference.run`.
        Idempotent dedup: identical bytes already uploaded in this organization return
        the existing scan with `duplicate: true` (HTTP 200) and enqueue no new job.
        Rate limit 20/hour/user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "202":
          description: Accepted for async processing.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [scan, job_id, job_status, duplicate, message]
                        properties:
                          scan: { $ref: "#/components/schemas/ScanDetail" }
                          job_id: { type: [string, "null"], format: uuid }
                          job_status:
                            type: [string, "null"]
                            enum: [QUEUED, RUNNING, RETRYING, COMPLETED, FAILED, CANCELLED]
                          duplicate: { type: boolean, default: false }
                          message: { type: [string, "null"] }
        "200":
          description: |
            Duplicate bytes already uploaded in this organization — the existing scan
            is returned with `duplicate: true` and no job is enqueued.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          scan: { $ref: "#/components/schemas/ScanDetail" }
                          job_id: { type: [string, "null"], format: uuid, default: null }
                          job_status: { type: [string, "null"], default: null }
                          duplicate: { type: boolean, enum: [true] }
                          message: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Error" }
        "413":
          description: SCAN_TOO_LARGE.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "415":
          description: UNSUPPORTED_FORMAT.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "422": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }

    get:
      tags: [Scans]
      operationId: listScans
      summary: Paginated scan history for the current user.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ScanStatus" }
      responses:
        "200":
          description: Paginated scan history.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/ScanSummary" }
                      meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/scans/{scan_id}:
    get:
      tags: [Scans]
      operationId: getScan
      summary: Scan status + result metadata (Redis-cached, PG fallback).
      security:
        - bearerAuth: []
      parameters:
        - name: scan_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Scan status and result metadata.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ScanDetail" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }

    delete:
      tags: [Scans]
      operationId: deleteScan
      summary: Soft delete a scan (sets deleted_at/deleted_by; in-flight job CANCELLED).
      security:
        - bearerAuth: []
      parameters:
        - name: scan_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Soft-deleted. }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...